+++ src/main/java/com/alibaba/dubbo/spring/boot/domain/DubboSpringBootStarterConstants.java
@@ -1,4 +1,4 @@
-package com.alibaba.dubbo.spring.boot.configuration;
+package com.alibaba.dubbo.spring.boot.domain;
 
 import lombok.experimental.UtilityClass;
